% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faulkner-package.R
\docType{package}
\name{faulkner-package}
\alias{faulkner-package}
\title{faulkner}
\description{
Palettes generated from original Henry Faulkner paintings.
}
\seealso{
\code{\link{faulkners}}, \code{\link{faulkner}}

\code{\link{imgpal}}, \code{\link{show_pal}}
}
