% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faulkners.R
\docType{data}
\name{faulkners}
\alias{faulkners}
\alias{faulkner}
\title{Faulkner palettes}
\format{
An object of class \code{list} of length 14.
}
\usage{
faulkners

faulkner(
  name,
  n = NULL,
  z = n,
  type = c("discrete", "continuous"),
  rev = FALSE
)
}
\arguments{
\item{name}{the palette name, one of \code{names(faulkners)} or an
unambiguous abbreviation}

\item{n}{the first \code{n} colors from the palette to use}

\item{z}{for \code{type = 'continuous'}, the number of colors to
interpolate from the sequence of \code{n} palette colors}

\item{type}{return a discrete or continuous (gradient) of colors}

\item{rev}{logical; if \code{TRUE}, the palette is reversed}
}
\description{
Color palettes stripped from the original paintings of Henry Faulkner.
}
\examples{
## some built-in palettes
names(faulkners)

op <- par(mfrow = n2mfrow(length(faulkners)))
for (ii in seq_along(faulkners))
  show_pal(structure(faulkners[[ii]], class = 'faulkner', name = names(faulkners)[ii]))
par(op)

## use or generate new palettes from existing
show_pal(faulkner('delphiniums1'))
show_pal(faulkner('delphiniums1', 4))
show_pal(faulkner('delphiniums1', 4, 100, type = 'continuous'))

## view palettes from other sources
# show_pal(nord::nord_palettes$afternoon_prarie)
show_pal(rainbow(8))

filled.contour(volcano, col = faulkner('delphiniums1', n = 4, 21, type = 'c'))
filled.contour(volcano, col = faulkner('delphiniums1', z = 21, type = 'c'))
filled.contour(volcano, col = faulkner('alice', 4, 21, type = 'c'))

}
\seealso{
\code{\link{imgpal}}; \code{\link{palette}}; \code{\link{colorRampPalette}};
\code{wesanderson::wes_palettes}; \code{nord::nord_palettes}
}
\keyword{datasets}
