% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_pal}
\alias{show_pal}
\title{Show palettes}
\usage{
show_pal(x, n = NULL, fullrange = FALSE, counts = inherits(x, "imgpal"))
}
\arguments{
\item{x}{one of 1) a \code{faulkners} name; 2) a vector of two or more
colors; 3) an \code{\link{imgpal}} object}

\item{n}{the first \code{n} colors will be shown}

\item{fullrange}{logical; for \code{\link{imgpal}} objects, if \code{TRUE},
the entire palette is shown; otherwise, only the unique colors (estimated
from ImageMagick) are shown}

\item{counts}{logical; for \code{\link{imgpal}} objects, if \code{TRUE},
the frequencies are shown for each color}
}
\description{
Show palettes as an image.
}
\examples{
show_pal(1:5)
show_pal(rainbow(8))
show_pal('alice')

go <- 'https://www.google.com/images/branding/googlelogo/2x/googlelogo_color_272x92dp.png'
ip <- imgpal(go)
show_pal(ip, n = 4)

}
